made changes requested
